Add flawed Gaussian process analysis
I have no confidence that this model actually fits, and it has huge
performance problems, because it's using each ride for a coefficient
instead of each departure time, and it's got the coefficients all screwy.
I need to scale the input data, and make sure that the kernel matrix is
only computed at distinct departure times.